Move things to a more natural place
